Moving Compare class inside utils module
